+++ src/main/java/pvpmode/PvPUtils.java
@@ -1,5 +1,7 @@
 package pvpmode;
 
+import java.io.*;
+import java.nio.file.*;
 import java.util.*;
 import java.util.function.Supplier;
 
@@ -10,6 +12,7 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.util.*;
 import net.minecraftforge.common.MinecraftForge;
+import pvpmode.overrides.EnumForcedPvPMode;
 
 public class PvPUtils
 {
@@ -122,15 +125,32 @@ public class PvPUtils
      */
     public static EnumPvPMode getPvPMode (EntityPlayer player)
     {
+        // Creative and flying players cannot do PvP
         if (isCreativeMode (player) || canFly (player))
-            return EnumPvPMode.OFF;// This is not really my (CraftedMods) style,
-                                   // but I'm doing this for performance reasons
-                                   // here, because the PvPData will only be
-                                   // loaded if required.
+            return EnumPvPMode.OFF;
 
         PvPData data = PvPUtils.getPvPData (player);
-        return data.getPvPWarmup () == 0 ? data.isPvPEnabled () ? EnumPvPMode.ON : EnumPvPMode.OFF
-            : EnumPvPMode.WARMUP;
+
+        EnumForcedPvPMode forcedPvPMode = data.getForcedPvPMode ();
+        if (!arePvPModeOverridesEnabled () || forcedPvPMode == EnumForcedPvPMode.UNDEFINED)
+        {
+            // No PvP mode overrides apply
+            if (data.getPvPWarmup () == 0)
+            {
+                // No warmup timer is running
+                return data.isPvPEnabled () ? EnumPvPMode.ON : EnumPvPMode.OFF;
+            }
+            else
+            {
+                // Warmup timer is running
+                return EnumPvPMode.WARMUP;
+            }
+        }
+        else
+        {
+            // PvP mode overrides apply
+            return forcedPvPMode.toPvPMode ();
+        }
     }
 
     /**
@@ -189,4 +209,49 @@ public class PvPUtils
         }
         return filledSlots;
     }
+
+    /**
+     * Returns whether the conditional PvP mode overrides are enabled.
+     */
+    public static boolean arePvPModeOverridesEnabled ()
+    {
+        return PvPMode.overrideCheckInterval != -1;
+    }
+
+    /**
+     * Returns whether the PvP mode for the player whose data are supplied is
+     * overridden.
+     */
+    public static boolean isPvPModeOverriddenForPlayer (PvPData data)
+    {
+        return arePvPModeOverridesEnabled () && data.getForcedPvPMode () != EnumForcedPvPMode.UNDEFINED;
+    }
+
+    /**
+     * Writes the contents of the supplied stream to the specified file.<br/>
+     * The file must exist on the filesystem.
+     * 
+     * @param stream
+     *            A supplier which creates the input stream
+     * @param file
+     *            The file where the data should be stored
+     * @throws IOException
+     *             If IO errors occur
+     */
+    public static void writeFromStreamToFile (Supplier<InputStream> stream, Path file) throws IOException
+    {
+        try (InputStream in = stream.get ();
+            InputStreamReader bridge = new InputStreamReader (in);
+            BufferedReader reader = new BufferedReader (bridge);
+            BufferedWriter writer = Files.newBufferedWriter (file))
+        {
+            String line = null;
+            while ( (line = reader.readLine ()) != null)
+            {
+                writer.write (line);
+                writer.newLine ();
+            }
+        }
+    }
+
 }