+++ src/main/java/pvpmode/command/PvPCommandConfig.java
@@ -60,6 +60,7 @@ public class PvPCommandConfig extends CommandBase
             Boolean.toString (PvPMode.partialInventoryLossEnabled));
         postConfigEntry (sender, "Armor item loss", Integer.toString (PvPMode.inventoryLossArmour) + " items");
         postConfigEntry (sender, "Hotbar item loss", Integer.toString (PvPMode.inventoryLossHotbar) + " items");
+        postConfigEntry (sender, "Override check interval", Integer.toString (PvPMode.overrideCheckInterval) + "s");
         PvPUtils.postChatLines (sender, EnumChatFormatting.GREEN, "---------------------------");
     }